pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "protencode"
version = "0.1.0"
description = "Different encoding/embedding types of mutations in amino acid sequences"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Chris Oldnall", email = "christopheraldous.oldnall@mdc-berlin.de" }
]

dependencies = [
    "pandas>=2.0",
    "numpy>=1.25",
    "tqdm>=4.67",
    "requests>=2.32",
    "biopython>=1.85",
    "matplotlib",
    "scipy",
    "pyyaml>=6.0",
    "torch==2.3.0",
    "transformers==4.39.3",
    "tokenizers==0.15.2",
    "safetensors==0.4.3",
    "huggingface-hub==0.23.2"
]

[project.scripts]
protencode = "protencode.cli:main"